Move JSON API into its own directory
